// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  agent: typeof agent;
  auth: typeof auth;
  chat: typeof chat;
  documents: typeof documents;
  http: typeof http;
  ingest: typeof ingest;
  jobs: typeof jobs;
  jobs_helpers: typeof jobs_helpers;
  ocr: typeof ocr;
  projects: typeof projects;
  ragClient: typeof ragClient;
  summarization: typeof summarization;
  uploads: typeof uploads;
  users: typeof users;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "internal">
>;